created new class and moved samples to a folder
